+++ src/config.rs
@@ -3,9 +3,8 @@ use colored::*;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use std::collections::HashMap;
-use std::fs;
-use std::io;
-use std::io::*;
+use std::io::{Read, Write};
+use std::{fs, io};
 
 /// App configuration, serialized as json in ~/.tod.cfg
 #[derive(Clone, Serialize, Deserialize, Eq, PartialEq, Debug)]
@@ -22,79 +21,51 @@ pub struct Config {
 }
 
 impl Config {
-    pub fn new(token: &str) -> Config {
+    pub fn new(token: &str) -> Result<Config, String> {
         let projects: HashMap<String, u32> = HashMap::new();
-        Config {
-            path: generate_path(),
+        Ok(Config {
+            path: generate_path()?,
             token: String::from(token),
             next_id: None,
             last_version_check: Some(time::today_string()),
             projects,
-        }
+        })
     }
 
-    pub fn create(self) -> Config {
+    pub fn create(self) -> Result<Config, String> {
         let json = json!(self).to_string();
-        let _bytes = fs::File::create(&self.path)
-            .expect("could not create file")
-            .write(json.as_bytes())
-            .expect("could not write to file");
-        self
+        let mut file = fs::File::create(&self.path).or(Err("Could not create file"))?;
+        file.write_all(json.as_bytes())
+            .or(Err("Could not write to file"))?;
+        Ok(self)
     }
 
     pub fn save(self) -> std::result::Result<String, String> {
         let json = json!(self).to_string();
 
-        let bytes = fs::OpenOptions::new()
+        fs::OpenOptions::new()
             .write(true)
             .read(true)
             .truncate(true)
             .open(&self.path)
-            .expect("Could not find config")
-            .write(json.as_bytes());
+            .or(Err("Could not find config"))?
+            .write_all(json.as_bytes())
+            .or(Err("Could not write to file"))?;
 
-        match bytes {
-            Ok(_) => Ok(String::from("✓")),
-            Err(_) => Err(String::from("Could not write to file")),
-        }
+        Ok(String::from("✓"))
     }
 
-    pub fn load(path: String) -> Config {
+    pub fn load(path: String) -> Result<Config, String> {
         let mut json = String::new();
 
         fs::File::open(&path)
-            .expect("Could not find file")
+            .or(Err("Could not find file"))?
             .read_to_string(&mut json)
-            .expect("Could not read to string");
+            .or(Err("Could not read to string"))?;
 
-        let config: Config = serde_json::from_str(&json).expect("Could not parse JSON");
-        let last_version = config.clone().last_version_check;
-        let new_config = Config {
-            last_version_check: Some(time::today_string()),
-            ..config
-        };
+        let config: Config = serde_json::from_str(&json).or(Err("Could not parse JSON"))?;
 
-        if last_version != Some(time::today_string()) {
-            match request::get_latest_version() {
-                Ok(version) if version.as_str() != VERSION => {
-                    println!(
-                        "Latest Tod version is {}, found {}. Run {} to update",
-                        version,
-                        VERSION,
-                        "cargo install tod".bright_cyan()
-                    );
-                    new_config.clone().save().unwrap();
-                }
-                Ok(_) => (),
-                Err(err) => println!(
-                    "{}, {:?}",
-                    "Could not fetch Tod version from Cargo.io".red(),
-                    err
-                ),
-            };
-        }
-
-        new_config
+        check_for_latest_version(config)
     }
 
     pub fn add_project(self, name: &str, number: u32) -> Config {
@@ -124,41 +95,72 @@ impl Config {
     }
 }
 
-pub fn get_or_create() -> Config {
-    let path: String = generate_path();
+pub fn get_or_create() -> Result<Config, String> {
+    let path: String = generate_path()?;
     let desc = "Please enter your Todoist API token from https://todoist.com/prefs/integrations ";
 
     match fs::File::open(&path) {
         Ok(_) => Config::load(path),
         Err(_) => {
-            let token = get_input(desc);
-            Config::new(&token).create()
+            let token = get_input(desc)?;
+            Config::new(&token)?.create()
         }
     }
 }
 
-pub fn generate_path() -> String {
-    let filename = if cfg!(test) { "test" } else { ".tod.cfg" };
+fn check_for_latest_version(config: Config) -> Result<Config, String> {
+    let last_version = config.clone().last_version_check;
+    let new_config = Config {
+        last_version_check: Some(time::today_string()),
+        ..config
+    };
 
-    let home_directory = dirs::home_dir().expect("could not get home directory");
-    let home_directory_str = home_directory
-        .to_str()
-        .expect("could not set home directory to str");
-    format!("{}/{}", home_directory_str, filename)
+    if last_version != Some(time::today_string()) {
+        match request::get_latest_version() {
+            Ok(version) if version.as_str() != VERSION => {
+                println!(
+                    "Latest Tod version is {}, found {}. Run {} to update",
+                    version,
+                    VERSION,
+                    "cargo install tod".bright_cyan()
+                );
+                new_config.clone().save().unwrap();
+            }
+            Ok(_) => (),
+            Err(err) => println!(
+                "{}, {:?}",
+                "Could not fetch Tod version from Cargo.io".red(),
+                err
+            ),
+        };
+    }
+
+    Ok(new_config)
 }
 
-pub fn get_input(desc: &str) -> String {
-    if cfg!(test) {
-        String::from("test")
-    } else {
-        let mut input = String::new();
-        println!("{}", desc);
-        io::stdin()
-            .read_line(&mut input)
-            .expect("error: unable to read user input");
+pub fn generate_path() -> Result<String, String> {
+    let filename = if cfg!(test) { "test" } else { ".tod.cfg" };
 
-        String::from(input.trim())
+    let home_directory = dirs::home_dir()
+        .ok_or_else(|| String::from("Could not find home directory"))?
+        .to_str()
+        .ok_or_else(|| String::from("Could not convert directory to string"))?
+        .to_owned();
+    Ok(format!("{}/{}", home_directory, filename))
+}
+
+pub fn get_input(desc: &str) -> Result<String, String> {
+    if cfg!(test) {
+        return Ok(String::from("test"));
     }
+
+    let mut input = String::new();
+    println!("{}", desc);
+    io::stdin()
+        .read_line(&mut input)
+        .or(Err("error: unable to read user input"))?;
+
+    Ok(String::from(input.trim()))
 }
 
 #[cfg(test)]
@@ -169,13 +171,13 @@ mod tests {
 
     #[test]
     fn new_should_generate_config() {
-        let config = Config::new("something");
+        let config = Config::new("something").unwrap();
         assert_eq!(config.token, String::from("something"));
     }
 
     #[test]
     fn set_and_clear_next_id_should_work() {
-        let config = Config::new("something");
+        let config = Config::new("something").unwrap();
         assert_eq!(config.next_id, None);
         let config = config.set_next_id(123123);
         assert_eq!(config.next_id, Some(123123));
@@ -185,13 +187,13 @@ mod tests {
 
     #[test]
     fn add_project_should_work() {
-        let config = Config::new("something");
+        let config = Config::new("something").unwrap();
         let mut projects: HashMap<String, u32> = HashMap::new();
         assert_eq!(
             config,
             Config {
                 token: String::from("something"),
-                path: generate_path(),
+                path: generate_path().unwrap(),
                 next_id: None,
                 last_version_check: Some(time::today_string()),
                 projects: projects.clone(),
@@ -203,7 +205,7 @@ mod tests {
             config,
             Config {
                 token: String::from("something"),
-                path: generate_path(),
+                path: generate_path().unwrap(),
                 next_id: None,
                 last_version_check: Some(time::today_string()),
                 projects,
@@ -218,7 +220,7 @@ mod tests {
         projects.insert(String::from("test2"), 4567);
         let config_with_two_projects = Config {
             token: String::from("something"),
-            path: generate_path(),
+            path: generate_path().unwrap(),
             next_id: None,
             last_version_check: None,
             projects: projects.clone(),
@@ -228,7 +230,7 @@ mod tests {
             config_with_two_projects,
             Config {
                 token: String::from("something"),
-                path: generate_path(),
+                path: generate_path().unwrap(),
                 next_id: None,
                 last_version_check: None,
                 projects: projects.clone(),
@@ -241,7 +243,7 @@ mod tests {
             config_with_one_project,
             Config {
                 token: String::from("something"),
-                path: generate_path(),
+                path: generate_path().unwrap(),
                 next_id: None,
                 last_version_check: None,
                 projects,
@@ -249,48 +251,42 @@ mod tests {
         );
     }
 
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-        use pretty_assertions::assert_eq;
+    #[test]
+    fn config_tests() {
+        // These need to be run sequentially as they write to the filesystem.
 
-        #[test]
-        fn config_tests() {
-            // These need to be run sequentially as they write to the filesystem.
+        // Save and load
+        // Build path
+        let home_directory = dirs::home_dir().expect("could not get home directory");
+        let home_directory_str = home_directory
+            .to_str()
+            .expect("could not set home directory to str");
+        let path = format!("{}/test", home_directory_str);
+        let _ = fs::remove_file(&path);
 
-            // Save and load
-            // Build path
-            let home_directory = dirs::home_dir().expect("could not get home directory");
-            let home_directory_str = home_directory
-                .to_str()
-                .expect("could not set home directory to str");
-            let path = format!("{}/test", home_directory_str);
-            let _ = fs::remove_file(&path);
+        // create and load
+        let new_config = Config::new("faketoken").unwrap();
+        let created_config = new_config.clone().create().unwrap();
+        assert_eq!(new_config, created_config);
+        let loaded_config = Config::load(path.clone()).unwrap();
+        assert_eq!(created_config, loaded_config);
 
-            // create and load
-            let new_config = Config::new("faketoken");
-            let created_config = new_config.clone().create();
-            assert_eq!(new_config, created_config);
-            let loaded_config = Config::load(path.clone());
-            assert_eq!(created_config, loaded_config);
+        // save and load
+        let different_new_config = Config::new("differenttoken").unwrap();
+        different_new_config.clone().save().unwrap();
+        let loaded_config = Config::load(path.clone()).unwrap();
+        assert_eq!(loaded_config, different_new_config);
+        assert_matches!(fs::remove_file(&path), Ok(_));
 
-            // save and load
-            let different_new_config = Config::new("differenttoken");
-            different_new_config.clone().save().unwrap();
-            let loaded_config = Config::load(path.clone());
-            assert_eq!(loaded_config, different_new_config);
-            assert_matches!(fs::remove_file(&path), Ok(_));
+        // get_or_create (create)
+        let config = get_or_create();
+        assert_eq!(config, Config::new("test"));
+        assert_matches!(fs::remove_file(&path), Ok(_));
 
-            // get_or_create (create)
-            let config = get_or_create();
-            assert_eq!(config, Config::new("test"));
-            assert_matches!(fs::remove_file(&path), Ok(_));
-
-            // get_or_create (load)
-            Config::new("alreadycreated").create();
-            let config = get_or_create();
-            assert_eq!(config, Config::new("alreadycreated"));
-            assert_matches!(fs::remove_file(&path), Ok(_));
-        }
+        // get_or_create (load)
+        Config::new("alreadycreated").unwrap().create().unwrap();
+        let config = get_or_create();
+        assert_eq!(config, Config::new("alreadycreated"));
+        assert_matches!(fs::remove_file(&path), Ok(_));
     }
 }