+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Check for latest version once per day and prompt to update with `cargo install tod`
+- Update dependencies
 
 # 0.2.2
 - Sort projects alphabetically when listing