+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Check for latest version once per day and prompt to update with `cargo install tod`
 - Update dependencies
+- Only show the time when a datetime is today
 
 # 0.2.2
 - Sort projects alphabetically when listing